+++ R/predict.R
@@ -5,17 +5,18 @@
 #' @param newdata a dfm on which prediction should be made.
 #' @param se_fit if `TRUE`, returns standard error of document scores.
 #' @param density if `TRUE`, returns frequency of polarity words in documents.
-#' @param divide if specified, polarity scores of words are dichotomized by the
-#'   percentile value.
+#' @param cut a vector of one or two percentile values to dichotomized polarty
+#'   scores of words. When two values are given, words between them receive zero
+#'   polarity.
 #' @param rescale if `TRUE`, normalizes polarity scores using `scale()`.
 #' @param min_n set the minimum number of polarity words in documents.
 #' @param ... not used
 #' @details Polarity scores of documents are the means of polarity scores of
 #'   words weighted by their frequency. When `se_fit = TRUE`, this function
 #'   returns the weighted means, their standard errors, and the number of
-#'   polarity words in the documents. When `rescale = TRUE`, it converts the
-#'   raw polarity scores to z sores for easier interpretation. When `rescale =
-#'   FALSE` and `divide` is used, polarity scores of documents are bounded by
+#'   polarity words in the documents. When `rescale = TRUE`, it converts the raw
+#'   polarity scores to z sores for easier interpretation. When `rescale =
+#'   FALSE` and `cut` is used, polarity scores of documents are bounded by
 #'   \[-0.5, 0.5\].
 #'
 #'   Documents tend to receive extreme polarity scores when they have only few
@@ -29,7 +30,7 @@
 #' @export
 predict.textmodel_lss <- function(object, newdata = NULL, se_fit = FALSE,
                                   density = FALSE, rescale = TRUE,
-                                  divide = NULL, min_n = 0L, ...){
+                                  cut = NULL, min_n = 0L, ...){
 
 
     (function(se.fit, recaling, ...) unused_dots(...))(...) # trap deprecated args
@@ -44,10 +45,9 @@ predict.textmodel_lss <- function(object, newdata = NULL, se_fit = FALSE,
     }
     min_n <- check_integer(min_n, min = 0)
 
-    if (!is.null(divide)) {
-        divide <- check_double(divide, min = 0, max = 1)
-        object$beta <- structure((object$beta > quantile(object$beta, divide)) - 0.5,
-                                 names = names(object$beta))
+    if (!is.null(cut)) {
+        cut <- check_double(cut, min = 0, max = 1, min_len = 1, max_len = 2)
+        object$beta <- cut_beta(object$beta, cut)
     }
 
     beta <- Matrix(object$beta, nrow = 1, sparse = TRUE,
@@ -100,3 +100,14 @@ predict.textmodel_lss <- function(object, newdata = NULL, se_fit = FALSE,
         return(result)
     }
 }
+
+cut_beta <- function(x, p = 0.5) {
+    q <- c(-Inf, quantile(x, p, na.rm = TRUE), Inf)
+    v <- as.integer(cut(x, q))
+    beta <- double(length(x))
+    beta[v == min(v)] <- -0.5
+    beta[v == max(v)] <- 0.5
+    names(beta) <- names(x)
+    return(beta)
+}
+