+++ NEWS.md
@@ -1,6 +1,6 @@
 # Changes
 
-* Add `divide` to `predict()` to produce bounded polarity scores of documents.
+* Add `cut` to `predict()` to produce bounded polarity scores of documents.
 * Rename "rescaling" to "rescale" for simplicity and consistency.
 
 # Changes in v1.2.0